% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{bigee_estimate}
\alias{bigee_estimate}
\title{Estimate BIGEE}
\usage{
bigee_estimate(
  bigeedata,
  bigeepriors = NULL,
  cores = getOption("mc.cores", default = parallel::detectCores()),
  chains = 3L,
  iter = 1000L,
  stanmodel = NULL,
  pars = NULL,
  include = FALSE,
  ...
)
}
\arguments{
\item{bigeedata}{A bigeedata object, as produced by \code{bigee_data()}}

\item{bigeepriors}{A bigeepriors object. If none is supplied, defaults are used
from calling \code{bigee_priors(bigeedata)} (with no other arguments).}

\item{cores}{Number of processing cores for running chains in parallel.
See \code{?rstan::sampling}. Defaults to \code{parallel::detectCores()}.}

\item{chains}{A positive integer specifying the number of Markov chains.
The default is 3.}

\item{iter}{Number of iterations per chain (including warmup). Defaults to 1000.}

\item{stanmodel}{A \code{stanmodel} object to use instead of one of the default
models.}

\item{pars}{(passed to \code{rstan::sampling()}) A vector of character strings specifying
parameters of interest to be returned in the stanfit object. If not specified,
a default parameter set is returned.}

\item{include}{(passed to \code{rstan::sampling()}) Defaults to FALSE, which
excludes parameters specified in \code{pars} from the returned model.}

\item{...}{Other arguments passed to rstan::sampling() for customizing the
Monte Carlo sampler}
}
\description{
Fits a BIGEE model of one of several variants using Hamiltonian Monte Carlo.
}
