% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bigee_priors}
\alias{bigee_priors}
\title{Establish prior hyperparameters for BIGEER estimation}
\usage{
bigee_priors(bigeedata, ...)
}
\arguments{
\item{bigeedata}{An object of class bigeedata, as returned by \code{bigee_data}}

\item{...}{Optional manually set parameters. Unquoted expressions are allowed,
e.g. \code{logk600_sd = cv2sigma(0.8)}. Additionally, any variables present in
\code{bigeedata} may be referenced, e.g. \code{lowerbound_logk600 = log(mean(Wobs)) + log(5)}}
}
\description{
Produces a bigeepriors object that can be passed to bigee_estimate function
}
