% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{bigee_valdata}
\alias{bigee_valdata}
\title{Create a data.frame for BAM validation}
\usage{
bigee_valdata(fit, k600obs)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bigee_estimate()}}

\item{qobs}{a vector of observed flow.}
}
\description{
Create a data.frame for BAM validation
}
