% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{bigee_validate}
\alias{bigee_validate}
\title{Calculate validation metrics and plots}
\usage{
bigee_validate(
  fit,
  k600obs,
  stats = c("RRMSE", "MRR", "SDRR", "NRMSE", "rBIAS", "CoV", "r2")
)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bigee_estimate()}}

\item{k600obs}{a vector of observed flow.}

\item{stats}{Which stats to include in the summary?}
}
\description{
Calculate validation metrics and plots
}
