% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIGEER-package.R
\docType{package}
\name{BIGEER-package}
\alias{BIGEER-package}
\alias{BIGEER}
\title{The 'BIGEER' package.}
\description{
Uses Bayesian Inference + SWOT observations to estimate normalized gas transfer velocity from large rivers
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
