% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{biger_priors}
\alias{biger_priors}
\title{Establish prior hyperparameters for BIGER estimation}
\usage{
biger_priors(bigerdata, ...)
}
\arguments{
\item{bigerdata}{An object of class bigerdata, as returned by \code{biger_data}}

\item{...}{Optional manually set parameters. Unquoted expressions are allowed,
e.g. \code{logk600_sd = cv2sigma(0.8)}. Additionally, any variables present in
\code{bigerdata} may be referenced, e.g. \code{lowerbound_logk600 = log(mean(Wobs)) + log(5)}}
}
\description{
Produces a bigerpriors object that can be passed to biger_estimate function
}
