% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bigge_check_args}
\alias{bigge_check_args}
\title{Performs the following checks:
- types:
    - everything else matrix
- dimensions:
    - all matrices have same dims}
\usage{
bigge_check_args(datalist)
}
\arguments{
\item{datalist}{A list of BIGGE data inputs}
}
\description{
Performs the following checks:
- types:
    - everything else matrix
- dimensions:
    - all matrices have same dims
}
