% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bigge_data}
\alias{bigge_data}
\title{Preprocess data for BIGGER estimation}
\usage{
bigge_data(w, s, dA, max_xs = 30L, seed = NULL)
}
\arguments{
\item{w}{Matrix (or data frame) of widths: time as columns, space as rows}

\item{s}{Matrix of slopes: time as columns, space as rows}

\item{dA}{Matrix of area above base area: time as columns, space as rows}

\item{max_xs}{Maximum number of cross-sections to allow in data. Used to reduce
sampling time. Defaults to 30.}

\item{seed}{RNG seed to use for sampling cross-sections, if nx > max_xs.}
}
\description{
Produces a bamdata object that can be passed to bam_estimate function
}
