% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{bigge_k600pred}
\alias{bigge_k600pred}
\title{k600 posterior mean and Bayesian credible interval.}
\usage{
bigge_k600pred(fit, chain = "all", conf.level = 0.95)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bigge_estimate()}}

\item{chain}{Either an integer specifying which chain(s) to extract statistics from,
or "all" (the default), in which case they are extracted from all chains.}

\item{conf.level}{A numeric value on (0, 1) specifying the size of the Bayesian 
credible interval. Default is 0.95.}
}
\description{
k600 posterior mean and Bayesian credible interval.
}
