% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bigge_priors}
\alias{bigge_priors}
\title{Establish prior hyperparameters for BIGGE estimation}
\usage{
bigge_priors(biggedata, ...)
}
\arguments{
\item{biggedata}{An object of class biggedata, as returned by \code{bigge_data}}

\item{...}{Optional manually set parameters. Unquoted expressions are allowed,
e.g. \code{logk600_sd = cv2sigma(0.8)}. Additionally, any variables present in
\code{biggedata} may be referenced, e.g. \code{lowerbound_logk600 = log(mean(Wobs)) + log(5)}}
}
\description{
Produces a biggepriors object that can be passed to bigge_estimate function
}
