% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{biker_check_args}
\alias{biker_check_args}
\title{Performs the following checks:
- types:
    - everything else matrix
- dimensions:
    - all matrices have same dims}
\usage{
biker_check_args(datalist)
}
\arguments{
\item{datalist}{A list of biker data inputs}
}
\description{
Performs the following checks:
- types:
    - everything else matrix
- dimensions:
    - all matrices have same dims
}
