% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{biker_data}
\alias{biker_data}
\title{Preprocess data for BIKER estimation}
\usage{
biker_data(w, s, dA, Kdesired, max_xs = 30L, seed = NULL)
}
\arguments{
\item{w}{Matrix (or data frame) of widths: time as columns, space as rows}

\item{s}{Matrix of slopes: time as columns, space as rows}

\item{dA}{Matrix of area above base area: time as columns, space as rows}

\item{Kdesired}{binary flag for which k to estimate: 0 for KO2, 1 for k600}

\item{max_xs}{Maximum number of cross-sections to allow in data. Used to reduce
sampling time. Defaults to 30.}

\item{seed}{RNG seed to use for sampling cross-sections, if nx > max_xs.}
}
\description{
Produces a bikerdata object that can be passed to biker_estimate function
}
