% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{biker_extract}
\alias{biker_extract}
\title{Extract posterior}
\usage{
biker_extract(fitmodel, CI = 0.95, chainExtract = "all")
}
\arguments{
\item{fitmodel}{A fitted BIKER stanmodel}

\item{CI}{A postive integer between 0 and 1 indicating the confidence interval to return with the estimate. Default is 0.95.}

\item{chainExtract}{Either an integer specifying which chain(s) to extract statistics from,
or "all" (the default), in which case they are extracted from all chains.}
}
\description{
Extracts posterior parameters from fit stan model
}
