% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dA_sigma_func}
\alias{dA_sigma_func}
\title{Assign measurement error for dA using river width and Durand et al. 2020}
\usage{
dA_sigma_func(Wobs)
}
\arguments{
\item{Wobs}{observed widths matrix}
}
\description{
Assign measurement error for dA using river width and Durand et al. 2020
}
