% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logA0}
\alias{estimate_logA0}
\title{logA0_hat prior estimation}
\usage{
estimate_logA0(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix}
}
\description{
Estimate median cross-sectional area prior hyperparameter using swot data
}
