% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logk}
\alias{estimate_logk}
\title{Estimate k_hat using biker data}
\usage{
estimate_logk(k600flag, Sobs)
}
\arguments{
\item{k600flag}{binary flag for model ran: 0 for KO2, 1 for k600}

\item{Sobs}{Observed S,as a space-down, time-across matrix}
}
\description{
Estimate k_hat using biker data
}
