% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logk600sd}
\alias{estimate_logk600sd}
\title{Estimate k600 sd prior using biker data}
\usage{
estimate_logk600sd(Wobs, Sobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix}
}
\description{
Estimate k600 sd prior using biker data
}
