% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logk_k600}
\alias{estimate_logk_k600}
\title{Estimate k_hat using biker data for k600 model}
\usage{
estimate_logk_k600(Sobs)
}
\arguments{
\item{Sobs}{Observed S,as a space-down, time-across matrix}
}
\description{
Estimate k_hat using biker data for k600 model
}
