% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logksd}
\alias{estimate_logksd}
\title{Estimate k sd prior using biker data}
\usage{
estimate_logksd(k600flag, Sobs)
}
\arguments{
\item{k600flag}{binary flag for model ran: 0 for KO2, 1 for k600}

\item{Sobs}{Observed S,as a space-down, time-across matrix}
}
\description{
Estimate k sd prior using biker data
}
