% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logksd_k600}
\alias{estimate_logksd_k600}
\title{Estimate k sd prior using biker data for k600 model}
\usage{
estimate_logksd_k600(Sobs)
}
\arguments{
\item{Sobs}{Observed S,as a space-down, time-across matrix}
}
\description{
Estimate k sd prior using biker data for k600 model
}
