% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logn}
\alias{estimate_logn}
\title{logn_hat prior estimation}
\usage{
estimate_logn(Wobs, Sobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix}

\item{Sobs}{Observed s, as a space-down, time-across matrix}
}
\description{
Estimate manning's n prior hyperparameter using swot data
}
