% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lognSD}
\alias{estimate_lognSD}
\title{logn_sd prior estimation}
\usage{
estimate_lognSD(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate manning's n SD prior hyperparameter using swot data
}
