% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundA0}
\alias{estimate_lowerboundA0}
\title{lowerbound_A0 prior estimation}
\usage{
estimate_lowerboundA0(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate median cross-sectional area lowerbound prior hyperparameter using swot data
}
