% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundlogk}
\alias{estimate_lowerboundlogk}
\title{Estimate k lowerbound prior using biker data}
\usage{
estimate_lowerboundlogk(k600flag)
}
\arguments{
\item{k600flag}{binary flag for model ran: 0 for KO2, 1 for k600}
}
\description{
Estimate k lowerbound prior using biker data
}
