% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundlogk600}
\alias{estimate_lowerboundlogk600}
\title{Estimate lowerbound k600 prior using bigge data}
\usage{
estimate_lowerboundlogk600(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix}
}
\description{
Estimate lowerbound k600 prior using bigge data
}
