% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundA0}
\alias{estimate_upperboundA0}
\title{Upperbound_A0 prior estimation}
\usage{
estimate_upperboundA0(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate median cross-sectional area upperbound prior hyperparameter using swot data
}
