% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundlogk600}
\alias{estimate_upperboundlogk600}
\title{Estimate upperbound k600 prior using bigee data}
\usage{
estimate_upperboundlogk600(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix}
}
\description{
Estimate upperbound k600 prior using bigee data
}
