% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gBiGER.R
\name{gBiGER}
\alias{gBiGER}
\title{BiGER algorithm for gene list aggregation.}
\usage{
gBiGER(
  r,
  n_r,
  n_u,
  mu = NULL,
  sigma2 = NULL,
  w = NULL,
  alpha = 1,
  beta = 1,
  save_chains = F,
  save_burnin = F,
  chains = 1,
  iter = 5000,
  burnin = 2500,
  verbose = -1
)
}
\arguments{
\item{r}{A integer rank matrix with rows as genes and columns as studies. All
ties are parametrized as ties with the `min` method. See [preprocess_genelist()]
for constructing this matrix.}

\item{n_r}{A vector containing the number of ranked items in each gene list.}

\item{n_u}{A vector containing the number of unranked items in each gene list.}

\item{mu}{The initialization for the global importance vector. Optional.}

\item{sigma2}{The initialization for study variance vector. Optional.}

\item{w}{The initialization for the latent weight matrix. Optional.}

\item{alpha}{The \eqn{\alpha} (shape) parameter for the Inverse Gamma prior
distribution. Defaults to `1`.}

\item{beta}{The \eqn{\beta} (rate) parameter for the Inverse Gamma prior
distribution. Defaults to `1`.}

\item{save_chains}{Whether to return the full chains of the MCMC run. If `FALSE`, only
posterior means of parameters are returned. For large studies, saving the full
chains can have a considerable RAM cost. Defaults to `FALSE`.}

\item{save_burnin}{Whether to save the burn-in period of the MCMC run, given
that `save_chains` is set to `TRUE`. This is used mainly for diagnostic
purposes. Defaults to `FALSE`.}

\item{chains}{The number of chains to run. Defaults to 1.}

\item{iter}{The number of MCMC iterations to run. Defaults to `5000`.}

\item{burnin}{The number of iterations to discard as burn-in. Defaults to `2500`.}

\item{verbose}{The number of iteration to print progress. No printouts are
shown when set to `-1` as the default.}
}
\value{
A nested list containing chains of a MCMC bBiGER run. Each chain
itself is a list with with posterior mean \eqn{\mu}, posterior mean of \eqn{\sigma^2},
and optionally the full chains of \eqn{\mu} and \eqn{\sigma^2}. All genes and
studies are reported in the order given in `r`. The full chains are reported
in matrices with rows as iterations and columns as parameters.
}
\description{
This is the BiGER algorithm using a Gibbs sampler. For boundary approximation,
see \link[BiGER]{bBiGER()}; for efficiency, use \link[BiGER]{vBiGER()} for a fast variational
inference-based method.
}
\seealso{
\link[BiGER]{bBiGER()} for bBiGER and \link[BiGER]{BiGER()} VI-based BiGER.
}
