% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_genelist}
\alias{preprocess_genelist}
\title{Preprocess gene lists}
\usage{
preprocess_genelist(
  genelist,
  ranking_status = NULL,
  n_r = NULL,
  n_u = NULL,
  na_as_ties = FALSE,
  tie_method = "min",
  standardize = FALSE
)
}
\arguments{
\item{genelist}{A list of of studies containing genes. For ranked studies,}

\item{ranking_status}{A boolean vector indicating whether each study
is ranked or unranked. This is unnecessary if `n_r` or `n_u` are provided,
or if the list has mixed gene lists.}

\item{n_r:}{The number of ranked items in each gene list. This is necessary
if there are mixed gene lists.}

\item{n_u:}{The number of unranked items in each gene list. This is necessary
if there are mixed gene lists.}

\item{na_as_ties:}{Whether missing genes from each study is considered as
bottom ties. If FALSE, they are considered as missing genes (NAs).}

\item{tie_method:}{The method used to rank tied genes. Acceptable methods
are: "min", "mean", "median", and "max".}

\item{standardize:}{Whether to perform min-max standardization to the rank
matrix.}
}
\value{
A list containing a numeric rank matrix, the number of ranked items
in each gene list, and the number of unranked items in each gene list.
}
\description{
This function is used to generate rank matrices used by the BiGER algorithms.
The input format is a list of ordered or unordered genes, each representing
results from a single study.
}
