% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_seurat}
\alias{preprocess_seurat}
\title{Preprocess differential expression results from Seurat.}
\usage{
preprocess_seurat(
  de_results,
  fold_change = c("all", "positive", "negative"),
  rank_by = c("p_val", "p_val_adj", "avg_log2FC")
)
}
\arguments{
\item{de_results:}{The differential analysis results from Seurat's `FindMarkers`
function. This must be a dataframe in the same format as the original outputs.}

\item{fold_change:}{Whether to use 'all', 'positive', or 'negative' fold
change.}

\item{rank_by:}{The column by which to rank the genes. It must be one of
'p_val', 'p_val_adj', or 'avg_log2FC'. If 'avg_log2FC' is chosen, the order
will be decreasing if `fold_change` is set to 'all' or 'positive'; otherwise,
the ranking is done using increasing values.}
}
\value{
A character vector containing ranked genes in order.
}
\description{
This function is used to generate a single gene list from a differential
analysis conducted using Seurat. Specifically, we expect results from a
the `FindMarkers` function, and we rank the genes to provide a gene list
using all provided criteria.
}
\details{
Note that this function supports only a single gene list, not a cohort of
studies with many gene lists. This is done because many study designs
exist, and the corresponding Seurat pipelines differ. If it is desired to
run comparisons and generate many gene lists from the same Seurat object,
it is recommended to conduct call this function multiple times with separate
differential analyses.
}
