% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_settings.R
\name{prior_settings}
\alias{prior_settings}
\title{Options manager for BIKER default prior settings for k600 model}
\usage{
prior_settings(..., .__defaults = FALSE, .__reset = FALSE)
}
\arguments{
\item{...}{(Optional) named settings to query or set.}

\item{.__defaults}{See \code{?settings::option_manager}}

\item{.__reset}{See \code{?settings::option_manager}}
}
\description{
Options manager for BIKER default prior settings for k600 model
}
