% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{r2}
\alias{r2}
\title{Coefficient of determination}
\usage{
r2(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
Coefficient of determination
}
