% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{biker_estimate}
\alias{biker_estimate}
\title{Estimate BIKER}
\usage{
biker_estimate(
  bikerdata,
  bikerpriors,
  cores = getOption("mc.cores", default = parallel::detectCores()),
  meas_error = FALSE,
  chains = 3L,
  iter = 1000L,
  CI = 0.95,
  chainExtract = "all",
  pars = NULL,
  include = FALSE,
  ...
)
}
\arguments{
\item{bikerdata}{A bikerdata object, as produced by \code{biker_data()}}

\item{bikerpriors}{A bikerpriors object.}

\item{cores}{Number of processing cores for running chains in parallel.
See \code{?rstan::sampling}. Defaults to \code{parallel::detectCores()}.}

\item{meas_error}{Should we run with latent variables accounting for uncertainty in SWOT measurements. LEAVE THIS OFF, IT IS IN ACTIVE DEVELOPMENT}

\item{chains}{A positive integer specifying the number of Markov chains. The default is 3.}

\item{iter}{Number of iterations per chain (including warmup). Defaults to 1000.}

\item{pars}{(passed to \code{rstan::sampling()}) A vector of character strings specifying parameters of interest to be returned in the stanfit object. If not specified, a default parameter set is returned.}

\item{include}{(passed to \code{rstan::sampling()}) Defaults to FALSE, which
excludes parameters specified in \code{pars} from the returned model.}

\item{...}{Other arguments passed to rstan::sampling() for customizing the
Monte Carlo sampler}

\item{CI:}{Confidence intervals, defaults to 95%, i.e.e 0.95}

\item{chainExtract:}{Which chains to use to construct the posterior k600. Defaults to all.}
}
\description{
Fits a BIKER model of one of several variants using Hamiltonian Monte Carlo.
}
