% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{biker_priors}
\alias{biker_priors}
\title{Establish prior hyperparameters for BIKER estimation}
\usage{
biker_priors(bikerdata, Kmodel = "ko2", ...)
}
\arguments{
\item{bikerdata}{An object of class bikerdata, as returned by \code{biker_data}}

\item{Kmodel}{Which k to estimate: 'ko2' for KO2, 'k600' for k600}

\item{...}{Optional manually set parameters. Unquoted expressions are allowed,
e.g. \code{logk_sd = cv2sigma(0.8)}. Additionally, any variables present in
\code{bikerdata} may be referenced, e.g. \code{lowerbound_logk = log(mean(Wobs)) + log(5)}}
}
\description{
Produces a bikerpriors object that can be passed to biker_estimate function
}
