% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logk}
\alias{estimate_logk}
\title{logk_hat prior estimation}
\usage{
estimate_logk(Sobs, priorQ)
}
\arguments{
\item{Sobs}{Observed S,as a space-down, time-across matrix}

\item{priorQ}{mean annual flow estimate, single number}
}
\description{
Estimate k_hat prior hyperparameter using swot data
}
