% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logk600}
\alias{estimate_logk600}
\title{Estimate k600_hat using biger data}
\usage{
estimate_logk600(Wobs, Sobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix}
}
\description{
Estimate k600_hat using biger data
}
